# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shelter Center API
  version: 1.0.0
  description: "유기동물 보호소 운영 API (Animals/Adoptions/Donations/Medical/Volunteers/Stats)"
servers:
  - url: http://localhost:8080
    description: local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [timestamp, path, status, code, message]
      properties:
        timestamp: { type: string, format: date-time }
        path: { type: string }
        status: { type: integer }
        code: { type: string }
        message: { type: string }
        details: { type: object, nullable: true }
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, sort]
      properties:
        content:
          type: array
          items: {}
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        sort: { type: string }
...
    MedicalRecord:
      type: object
      properties:
        id: { type: integer }
        animalId: { type: integer }
        type: { type: string, example: SURGERY }
        description: { type: string, nullable: true }
        performedAt: { type: string, format: date-time }
        cost: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /api/medical/animals/{id}/medical-records:
    post:
      tags: [Medical]
      summary: Create medical record for animal (sub-resource)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, description: "animalId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, performedAt]
              properties:
                type: { type: string, maxLength: 30 }
                description: { type: string, nullable: true }
                performedAt: { type: string, format: date-time }
                cost: { type: number, nullable: true }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MedicalRecord" }
        "404":
          description: animal not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    get:
      tags: [Medical]
      summary: List medical records by animal (paged)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, description: "animalId" }
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - in: query
          name: sort
          schema: { type: string, default: "performedAt,DESC" }
      responses:
        "200":
          description: paged medical records for animal
          content:
            application/json:
              schema:
                allOf: [{ $ref: "#/components/schemas/PageResponse" }]
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/MedicalRecord" }
        "404":
          description: animal not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/medical/medical-records:
    get:
      tags: [Medical]
      summary: List medical records (filter by type/date, paged)
      parameters:
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - in: query
          name: sort
          schema: { type: string, default: "performedAt,DESC" }
      responses:
        "200":
          description: paged medical records
          content:
            application/json:
              schema:
                allOf: [{ $ref: "#/components/schemas/PageResponse" }]
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/MedicalRecord" }
        "400":
          description: invalid query
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/medical/medical-records/{recordId}:
    get:
      tags: [Medical]
      summary: Get medical record by id
      parameters:
        - in: path
          name: recordId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MedicalRecord" }
        "404":
          description: record not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    patch:
      tags: [Medical]
      summary: Update medical record (partial)
      parameters:
        - in: path
          name: recordId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string }
                description: { type: string, nullable: true }
                performedAt: { type: string, format: date-time }
                cost: { type: number, nullable: true }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MedicalRecord" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: record not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    delete:
      tags: [Medical]
      summary: Delete medical record (ADMIN planned)
      parameters:
        - in: path
          name: recordId
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: deleted
        "404":
          description: record not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/medical/animals/{id}/medical-summary:
    get:
      tags: [Medical]
      summary: Recent medical summary for animal
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, description: "animalId" }
        - in: query
          name: limit
          schema: { type: integer, default: 5, minimum: 1, maximum: 20 }
      responses:
        "200":
          description: recent summary rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  animalId: { type: integer }
                  limit: { type: integer }
                  recent:
                    type: array
                    items: { $ref: "#/components/schemas/MedicalRecord" }
        "404":
          description: animal not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  ########################################################
  # Volunteers (mounted: /api/volunteers)  ※ 내부가 /volunteers..., /stats...
  ########################################################
...